+++ main/src/main/scala/org/clulab/dynet/CoNLLSRLToMetal.scala
@@ -363,7 +363,7 @@ object CoNLLSRLToMetal {
 
   def main(args: Array[String]): Unit = {
     assert(args.length == 2)
-    Utils.initializeDyNet()
+    Utils.initializeDyNet(train = false)
 
     val file = new File(args(0))
     val reader = new CoNLLSRLToMetal