+++ main/src/main/scala/org/clulab/dynet/FinalLayer.scala
@@ -2,7 +2,7 @@ package org.clulab.dynet
 
 import edu.cmu.dynet.{Expression, ExpressionVector}
 
-trait FinalLayer extends Saveable {
+trait FinalLayer extends Saveable with Cloneable {
   def forward(inputExpressions: ExpressionVector,
               headPositionsOpt: Option[IndexedSeq[Int]],
               doDropout: Boolean): ExpressionVector
@@ -15,4 +15,6 @@ trait FinalLayer extends Saveable {
   def inference(emissionScores: Array[Array[Float]]): IndexedSeq[String]
 
   def inferenceWithScores(emissionScores: Array[Array[Float]]): IndexedSeq[IndexedSeq[(String, Float)]]
+
+  override def clone(): FinalLayer = ???
 }