+++ main/src/main/scala/org/clulab/dynet/ViterbiForwardLayer.scala
@@ -6,17 +6,19 @@ import edu.cmu.dynet.{Dim, Expression, ExpressionVector, FloatVector, LookupPara
 import org.clulab.dynet.Utils.{ByLineFloatBuilder, ByLineIntBuilder, ByLineStringMapBuilder, LOG_MIN_VALUE, START_TAG, STOP_TAG, fromIndexToString, mkTransitionMatrix, save}
 import ForwardLayer._
 
-class ViterbiForwardLayer(parameters:ParameterCollection,
-                          inputSize: Int,
-                          isDual: Boolean,
-                          t2i: Map[String, Int],
-                          i2t: Array[String],
-                          H: Parameter,
-                          val T: LookupParameter, // transition matrix for Viterbi; T[i][j] = transition *to* i *from* j, one per task
-                          rootParam: Parameter, 
-                          nonlinearity: Int,
-                          dropoutProb: Float)
-  extends ForwardLayer(parameters, inputSize, isDual, t2i, i2t, H, rootParam, nonlinearity, dropoutProb) {
+case class ViterbiForwardLayer(override val parameters:ParameterCollection,
+                               override val inputSize: Int,
+                               override val isDual: Boolean,
+                               override val t2i: Map[String, Int],
+                               override val i2t: Array[String],
+                               override val H: Parameter,
+                               T: LookupParameter, // transition matrix for Viterbi; T[i][j] = transition *to* i *from* j, one per task
+                               override val rootParam: Parameter,
+                               override val nonlinearity: Int,
+                               override val dropoutProb: Float)
+  extends ForwardLayer {
+
+  override def clone(): ViterbiForwardLayer = this
 
   // call this *before* training a model, but not on a saved model
   def initializeTransitions(): Unit = {