+++ main/src/main/scala/org/clulab/sequences/LexiconNERShell.scala
@@ -21,7 +21,7 @@ class LexiconNERShell(val lexiconNer: LexiconNER) extends Shell {
 }
 
 object LexiconNERShell extends App {
-  Utils.initializeDyNet()
+  Utils.initializeDyNet(train = false)
   val kb = args(0) // pass the KB file name as the first argument
   val lexiconNer = LexiconNER(Seq(kb))
   val shell = new LexiconNERShell(lexiconNer)