+++ main/src/test/scala/org/clulab/processors/TestUniversalEnhancedDependencies.scala
@@ -6,7 +6,7 @@ import org.scalatest.{FlatSpec, Matchers}
 
 class TestUniversalEnhancedDependencies extends FlatSpec with Matchers {
     val proc = {
-    Utils.initializeDyNet()
+    Utils.initializeDyNet(train = false)
     new CluProcessor()
   }